+++ src/deprecated/routes/RequestsRoute/RequestsRoute.test.js
@@ -26,6 +26,7 @@ import {
   Checkbox,
   exportToCsv,
   NoValue,
+  dayjs,
 } from '@folio/stripes/components';
 import { effectiveCallNumber } from '@folio/stripes/util';
 
@@ -533,6 +534,14 @@ describe('RequestsRoute', () => {
     fulfillmentPreference: 'Hold Shelf',
   };
   const sendCallout = jest.fn();
+
+  dayjs.mockImplementation(() => ({
+    format: jest.fn(),
+    isSameOrBefore: jest.fn(),
+    tz: () => ({
+      toISOString: jest.fn(),
+    }),
+  }));
 
   const renderComponent = (props = defaultProps) => {
     const { rerender } = render(