+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 13.0.0 IN PROGRESS
 
 * *BREAKING* Use `convertToSlipData` and supporting functions from `stripes-util`. Refs UIREQ-1263.
+* Replace moment with day.js. Refs UIREQ-1291.
 
 ## [12.0.1] (https://github.com/folio-org/ui-requests/tree/v12.0.1) (2025-03-21)
 [Full Changelog](https://github.com/folio-org/ui-requests/compare/v12.0.0...v12.0.1)